requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cosmicpy==1.0.2
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6